pipeline/stream/requirements.txt
boto3
kafka-python
lycon
numpy
pillow
imageio
tqdm
